requirements.txt
aniso8601==8.0.0
bs4==0.0.1
click==7.1.2
Flask==1.1.2
Flask-Login==0.4.0
Flask-WTF==0.14.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
flask-sqlalchemy-bind==1.0.4
Flask-Bcrypt==0.7.1
pandas==1.0.5
itsdangerous==1.1.0
Jinja2==2.11.2
google-api-python-client==2.0.2
google-auth==1.28.0
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
xmltodict==0.10.2
werkzeug==0.16.1
advertools==0.13.2
collection==0.1.6
termcolor==2.3.0
gunicorn==20.0.4
requests==2.25.1
google==2.0.3
google-search==1.0.2
googlesearch-python==1.2.3
google-auth-httplib2==0.1.0